Fix warnings when compiling with clang 8
Because RapidCheck builds with the werror flag, warnings appearing with newer compilers might cause some issues. It is possible to hit this when building with clang 8. The error -Wself-assign-overloaded is
reported in some of the test cases for the project, which actually  test for self-assignment. To continue building with no warnings, we can add a small indirection where we just take a local reference and perform the self-assignment using it. This reduces the need to use pragmas or the need to modify the with warning flags for the test files.

Additionally, a deprecation warning is reported in ExecCommandsTests.cpp because a pragma that disables the warning locally is used only when compiling with gcc. That check should be extended to also handle clang since the same code works for clang as well.