Apply a linter to the documentation
The same way we have errors for our compiled code, it is nice to also apply the same techniques to the documentation. As such,
I ran markdownlint (https://github.com/DavidAnson/markdownlint) to the documentation files. Most of the suggestions were trivial but
actually, make the file structure a bit more reasonable to maintain. In addition to this, there were a couple of typos that I fixed while going through the documentation.

In general, I think this makes the "dev" part of the documentation a bit more consistent and nicer.