Improve CMake installation experience
The installation of RapidCheck and it subsequent usage as an external package (for example, through find_package)
which is a typical approach when using a package manager is not optimal. Firstly, the include directory is not very
well specified as one has to go one directory back if the installation include directories of the rapidcheck CMake
target should be used in the config file. Moreover, none of the integrations are installed which means that RapidCheck
cannot optimally be used as an external library. For example, one cannot use the Catch integration easily without
the need to manually copy around files.

With this commit, RapidCheck can more easily be integrated with numerous CMake-based (and even other) package managers
which are gaining popularity in C++. For examples, the addition of RapidCheck to VCPKG should be trivial after the
bellow changes are applied.

This commit does the following:
 - Fixes the installation include path of the rapidcheck target.
 - Adds installations of all integrations.
 - Adds the option RC_INSTALL_ALL_EXTRAS to make it simpler to add all the integrations without the need to
   define any of the targets found in the folder ext.
 - Some white space and reordering changes