Revert "Add an explicit option for building RapidCheck as either a static or a shared libary"
This is not the most standard way to choose whether a library should be dynamic or static. It is better to use BUILD_SHARED_LIBS as that is built into CMake. An extranal project including RapidCheck can simply set the variable BUILD_SHARED_LIBS externally and the library will build as normal.

This reverts commit a195151261e3480fd22a8878097213054181c7d8.